Extract TokenFactory module for multitest.
