+++ contracts/vault/Cargo.toml
@@ -39,4 +39,3 @@ anyhow        = { workspace = true }
 cw-multi-test = { workspace = true, features = ["cosmwasm_1_1"] }
 mars-testing  = { workspace = true }
 test-case     = { workspace = true }
-regex         = { version = "1.7.3" }
