+++ packages/testing/Cargo.toml
@@ -42,6 +42,9 @@ mars-swapper-astroport         = { workspace = true }
 mars-types                     = { workspace = true }
 prost                          = { workspace = true }
 pyth-sdk-cw                    = { workspace = true }
+regex                          = { workspace = true }
+schemars                       = { workspace = true }
+serde                          = { workspace = true }
 
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies]
 cw-it         = { workspace = true, features = ["multi-test"] }
