+++ packages/testing/src/lib.rs
@@ -11,6 +11,7 @@ mod incentives_querier;
 mod mars_mock_querier;
 mod mock_address_provider;
 mod mocks;
+pub mod multitest;
 mod oracle_querier;
 mod osmosis_querier;
 mod params_querier;
