+++ packages/testing/src/multitest/modules/mod.rs
@@ -1,0 +1,1 @@
+pub mod token_factory;
